moved create_img_frame to utils, fix skipping pressed key bug
